정적 분석 후 수정

+++ src/main/java/com/dasibom/practice/service/DiaryServiceImpl.java
@@ -114,7 +114,6 @@ public class DiaryServiceImpl implements DiaryService {
         }
 
         Pet pet = findPet(updateRequestDto.getPet(), user);
-//        Pet pet = updatePet(updateRequestDto, user);
         List<DiaryStamp> newDiaryStamps = updateStamp(updateRequestDto, diary);
         diary.updateDiary(updateRequestDto.getTitle(), updateRequestDto.getContent(), newDiaryStamps, pet);
     }
@@ -187,11 +186,9 @@ public class DiaryServiceImpl implements DiaryService {
         }
 
         // 기존 스탬프 제거
-        if (updateRequestDto.getStamps() != null) {
-            if (!oldDiaryStamps.isEmpty()) {
-                DiaryStamp.removeDiaryStamp(oldDiaryStamps);
-                stampRepository.deleteAll(oldStamps);
-            }
+        if (updateRequestDto.getStamps() != null && !oldDiaryStamps.isEmpty()) {
+            DiaryStamp.removeDiaryStamp(oldDiaryStamps);
+            stampRepository.deleteAll(oldStamps);
         }
 
         // 새로운 스탬프 생성
